move to connect directory
