+++ core/src/test/resources/META-INF/services/com.datastrato.gravitino.authorization.AuthorizationProvider
@@ -2,5 +2,5 @@
 # Copyright 2024 Datastrato Pvt Ltd.
 # This software is licensed under the Apache License version 2.
 #
-com.datastrato.gravitino.authorization.chain.authorization1.TestAuthorization1
-com.datastrato.gravitino.authorization.chain.authorization2.TestAuthorization2
+com.datastrato.gravitino.connector.authorization.authorization1.TestAuthorization1
+com.datastrato.gravitino.connector.authorization.authorization2.TestAuthorization2
